Extract common code to Prompters::BasePrompter
Also move QuestionPrompter to Prompters::ScriptCreationPrompter.
